Add files via upload
This file contains details of new customers for who we have to test our term deposit campaign will be successful or not. Also the files consists of attribute weightage of each feature and thus the total score for every customer based on those feature value. We then arrange the customers based on highest score to lowest.